resources/fastAPI_MQTT_WS/requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
fastapi==0.74.0
fastapi-mqtt==0.4.0
gmqtt==0.6.11
h11==0.13.0
idna==3.3
importlib-metadata==4.11.1
jsonpickle==2.1.0
paho-mqtt==1.6.1
pydantic==1.9.0
python-nmap==0.7.1
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.5
zipp==3.7.0
